% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sexbiaseddisp.R
\name{AIc}
\alias{AIc}
\title{Calculates corrected Assignment Index}
\usage{
AIc(dat)
}
\arguments{
\item{dat}{a data frane with nlocs+1 columns,}
}
\value{
aic  The corrected assignment index of each individual
}
\description{
Calculates corrected Assignment Index as described in Goudet etal. (2002)
}
\details{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
\references{
Goudet J, Perrin N, Waser P (2002) Tests for sex-biased dispersal
 using bi-parentally inherited genetic markers 11, 1103:1114
  \url{http://onlinelibrary.wiley.com/doi/10.1046/j.1365-294X.2002.01496.x/full}
}

