% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.betai.R
\name{betai.unw}
\alias{betai.unw}
\title{Estimation of \eqn{\beta} per population}
\usage{
betai.unw(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
Betai a table (nloc*npop) of individual loci \eqn{\beta_i}

Betai.ovl a vector of population's \eqn{\beta_i}, average over loci

BetaW a vector of locus \eqn{\beta_w}, where average is taken for each
locus over populations

Hi a table (nloc*npop) of within population diversities

Hb a vector (nloc) of between populations diversities
}
\description{
Estimates \eqn{\beta} per population, using unweighted
allele frequencies
}
\details{
Beware, only valid for non inbred diploids.
}
\examples{
dat<-sim.genot(size=100,N=c(100,1000,10000),nbloc=50,nbal=10)
 betai.unw(dat)$Betai
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}

