% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont.isl}
\alias{cont.isl}
\title{A genetic dataset from a diploid organism in a continent-island model}
\format{An object of class \code{data.frame} with 150 rows and 6 columns.}
\source{
generated with function sim.genot()
}
\usage{
cont.isl
}
\description{
A simple diploid dataset, with allele encoded as one digit number.  
Up to 4 alleles per locus
}
\details{
\describe{
  \item{Pop}{Population identifier, from 1 to 3}
  \item{loc.1}{genotype at loc.1}
  \item{loc.2}{genotype at loc.2}
  \item{loc.3}{genotype at loc.3}
  \item{loc.4}{genotype at loc.4}
  \item{loc.5}{genotype at loc.5}
 
  ...
}
}
\examples{
data(cont.isl)
allele.count(cont.isl)

}
\keyword{datasets}
