% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{diploid}
\docType{data}
\alias{diploid}
\title{A genetic dataset from a diploid organism}
\format{An object of class \code{data.frame} with 44 rows and 6 columns.}
\source{
Given in Weir, B.S. Genetic Data Analysis. Sinauer
}
\usage{
data(diploid)
}
\description{
A simple diploid dataset, with allele encoded as one digit number
}
\details{
\describe{
  \item{Pop}{Population identifier, from 1 to 6}
  \item{loc-1}{genotype at loc-1 (only allele 4 present)}
  \item{loc-2}{genotype at loc-1 (alleles 3 and 4)}
  \item{loc-3}{genotype at loc-1 (alleles 2, 3 and 4)}
  \item{loc-4}{genotype at loc-1 (alleles 1, 2, 3 and 4)}
  \item{loc-5}{genotype at loc-1 (only allele 4)}
 
  ...
}
}
\examples{
data(diploid)
basic.stats(diploid)

}
\keyword{datasets}
