\name{eucl.dist.trait}
\alias{eucl.dist.trait}
\title{calculates euclidian distance among populations for a trait}
\description{calculates euclidian distance among populations for a trait}
\usage{eucl.dist.trait(data)}
\arguments{
\item{data}{a 2 columns data frame, the first columen identifying the population of origin, the second teh value for the trait}
}

\value{
\item{}{a vector of euclidian distances 2-1, 3-1, 3-2 etc}
} 
%\references{}
\author{Jerome Goudet \email{jerome.goudet@unil.ch}}

%\seealso{\code{\link{}}.}
%\examples{}
