% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdosage.R
\name{fis.dosage}
\alias{fis.dosage}
\title{Reports population Fiss from dosage data}
\usage{
fis.dosage(dos, pop, matching = FALSE)
}
\arguments{
\item{dos}{either a matrix with snps columns and individuals in rows containing allelic dosage (number [0,1 or 2] of alternate alleles); 
or a square matrix with as many rows and columns as the number of individuals and containing the proportion of matching alleles}

\item{pop}{a vector containing the identifier of the population to which the individual in the corresponding row belongs}

\item{matching}{logical:TRUE if dos is a square matrix of allelic matching; FALSE otherwise}
}
\value{
population specific and overall Fis.
}
\description{
Reports population Fiss from dosage data. see \link{fs.dosage}
}
\details{
@usage fis.dosage(dos, pop, matching = FALSE)
}
