% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksGRM.R
\name{fst.dosage}
\alias{fst.dosage}
\title{Calculates population specific \eqn{F_{ST}} from dosage data}
\usage{
fst.dosage(dat,pop,matching=FALSE)
}
\arguments{
\item{dat}{either a niXnl dosage matrix of 0,1 and 2s containing the 
number of derived alleles; or a niXni matrix of matching proportions}

\item{pop}{a vector of length ni containing the identifier of the population
to which individuals belong}

\item{matching}{a logical element, whether \code{dat} is dosage (niXnl, default) or 
matching (niXni)  matrix.}
}
\description{
Calculates population specific \eqn{F_{ST}} from dosage data
}
\note{
Does not assume Hardy-Weinberg equilibrium. If the numbers of samples per population 
are identical, fst.dosage gives the same results as \code{\link{wc}} for overall \eqn{F_{ST}}
}
\examples{
\dontrun{ 
dos<-matrix(sample(0:2,replace=TRUE,size=1000),nrow=10) #dosage matrix for 10 inds at 100 loci
fst.dosage(dos,rep(1:2,each=5)) 

ks<-beta.dosage(dos,Mb=TRUE) # kinship matrix
fst.dosage(ks$betas*(1-ks$MB)+ks$MB,pop=rep(1:2,each=5),matching=TRUE)

dat<-sim.genot(size=100,nbal=2,nbloc=200)
c(betas(dat)$betaiovl,betas(dat)$betaW)
dos<-as.matrix((dat[,-1]\%/\%10-1)+(dat[,-1]\%\%10-1))
fst.dosage(dos,pop=rep(1:3,each=100))
wc(dat)$FST # same as fst.dosage overall if sampling is balanced 
}
}
\references{
\href{http://www.genetics.org/content/206/4/2085}{Weir, BS and Goudet J. 2017} A Unified Characterization 
of Population Structure and Relatedness. Genetics (2017) 206:2085
}
\seealso{
The \code{\link{betas}} function estimates population specific \eqn{F_{ST}} from non-dosage data, 
assuming random mating
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
