% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{genind2hierfstat}
\alias{genind2hierfstat}
\title{Converts genind objects from adegenet into a hierfstat data frame}
\usage{
genind2hierfstat(dat,pop=NULL)
}
\arguments{
\item{dat}{a genind object}

\item{pop}{a vector containing the population to which each individual belongs}
}
\value{
a dataframe with nloci+1 columns and ninds rows.  The first column
contains the population identifier, the following the genotypes at each locus
}
\description{
Converts genind objects from adegenet into a hierfstat data frame
}

