% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksGRM.R
\name{grm2kinship}
\alias{grm2kinship}
\title{Converts a Genetic Relationship Matrix (GRM) to a kinship matrix}
\usage{
grm2kinship(x)
}
\arguments{
\item{x}{a square (GRM) matrix}
}
\value{
a kinship matrix
}
\description{
Converts a Genetic Relationship Matrix (GRM) to a kinship matrix
}
\details{
\eqn{k[ii]=x[ii]-1; k[ij]=x[ij]/2}
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
