\name{gtrunchier}
\docType{data}
\alias{gtrunchier}
\title{Genotypes at 6 microsatellite loci of Galba truncatula from different patches in Western Switzerland}
\description{Data set consisting of the microsatellite genotypes of 370 Galba truncatula, a tiny freshwater snail, collecting from different localities and several patches within localities in Western Switzerland.}
\usage{
data(gtrunchier)
}
\value{
\item{Locality}{Identifier of the locality of origin}
\item{Patch}{Identifier of the patch of origin}
\item{L21.V}{Genotype at locus L21.V. For instance the first individual carries allele 2 and 2 at this locus \preformatted{gtrunchier\$L21.V[1]}}
\item{L37.J}{Genotype at locus L37.J}
\item{L20.B}{Genotype at locus L20.B}
\item{L29.V}{Genotype at locus L29.V}
\item{L36.B}{Genotype at locus L36.B}
\item{L16.J}{Genotype at locus L16.J}

}

\references{
Trouve S., L. Degen et al. (2000) Microsatellites in the hermaphroditic snail, Lymnaea truncatula, intermediate host of the liver fluke, Fasciola hepatica.Molecular Ecology 9: 1662-1664.

Trouve S., Degen L. and Goudet J. (2005) Ecological components and evolution of selfing in the freshwater snail Galba truncatula. Journal of Evolutionary Biology. 18, 358-370
}
\keyword{datasets}

