% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.genind}
\alias{is.genind}
\title{Test if object is of class \code{\link[adegenet]{genind}}}
\usage{
is.genind(dat)
}
\arguments{
\item{dat}{an object}
}
\value{
TRUE or FALSE
}
\description{
Test if the argument to the function is of class 
\code{\link[adegenet]{genind}}, a class from the \code{\link[adegenet]{adegenet}} library)
}
