% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksGRM.R
\name{kinship2dist}
\alias{kinship2dist}
\title{Converts a kinship matrix to a distance matrix}
\usage{
kinship2dist(x)
}
\arguments{
\item{x}{A square matrix containg kinship coefficients}
}
\description{
Converts a kinship matrix to a distance matrix
}
\details{
\eqn{D_{ii}=0, D_{ij}=\frac{1-(x-min(x))}{(1-min(x))}}
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
