% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmsb.R
\name{ms2bed}
\alias{ms2bed}
\title{Import the output of the \code{ms} program in a \code{BED} object}
\usage{
ms2bed(fname)
}
\arguments{
\item{fname}{the name of the text file containing \code{ms} output}
}
\value{
a bed object
}
\description{
Import the output of the \code{ms} program into a \code{BED} object, as defined in the 
\href{https://cran.r-project.org/web/packages/gaston/gaston.pdf}{gaston} package
}
