% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmsb.R
\name{ms2dos}
\alias{ms2dos}
\title{Import \code{ms} output}
\usage{
ms2dos(fname)
}
\arguments{
\item{fname}{a text file containing the output of the \code{ms program}}
}
\value{
alldat a matrix with as many row as (haploid) individuals and as many columns as SNPs

bim a data frame with two components
 chr contains the chromosome (replicate) id; pos contains the SNPs posoition on the chromosome
}
\description{
Import the output of the \code{ms} program into suitable format for further manipulation
}
