% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdosage.R
\name{plot.fs.dosage}
\alias{plot.fs.dosage}
\title{Plot function for fs.dosage}
\usage{
plot.fs.dosage(x,...)
}
\arguments{
\item{x}{a fs.dosage object}

\item{...}{further arguments to pass}
}
\description{
Plot function for fs.dosage
}
