\name{pp.sigma.loc}
\alias{pp.sigma.loc}
\title{wrapper to return per locus variance components}
\description{wrapper to return per locus variance components between pairs of samples x & y}
\usage{pp.sigma.loc(x,y,dat=dat,diploid=TRUE,...)}
\arguments{
\item{x,y}{samples 1 and 2}
\item{dat}{a genetic data set}
\item{diploid}{whether dats are diploid}
\item{...}{further arguments to pass to the function}
}

\value{
\item{sigma.loc}{variance components per locus}
} 
%\references{}
\author{Jerome Goudet \email{jerome.goudet@unil.ch}}

%\seealso{\code{\link{}}.}
%\examples{}
\keyword{univar}
