% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdosage.R
\name{print.fs.dosage}
\alias{print.fs.dosage}
\title{Print function for fs.dosage}
\usage{
print.fs.dosage(x,...)
}
\arguments{
\item{x}{a fs.dosage object}

\item{...}{further arguments to pass}
}
\description{
Print function for fs.dosage
}
