\name{print.pp.fst}
\alias{print.pp.fst}
\title{print function for pp.fst}
\description{print function for pp.fst}
\usage{
\method{print}{pp.fst}(x,...)}
\arguments{
\item{x}{an object of class \kbd{pp.fst}}
\item{...}{further arguments to pass to the function}
}

\author{Jerome Goudet \email{jerome.goudet@unil.ch}}

\keyword{univar}
