% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readms.R
\name{read.ms.snps}
\alias{read.ms.snps}
\title{Read data generated by Hudson ms program}
\usage{
ms.read.snps(fname)
}
\arguments{
\item{fname}{ms data file name}
}
\value{
alldat  a data frame with nloc+1 columns, the first being the population
 to which the individual belongs and the next being the genotypes, one column per locus; 
 and ninds rows
}
\description{
Read data generated by \href{http://home.uchicago.edu/rhudson1/source/mksamples.html}{Hudson ms} program
}
\details{
Each site is read as a SNP, with the ancestral allele encoded as 0 and the alternate 
allele encoded as 1.  If the ms output file contains several replicates, 
the different replicates will be collated together.
}
\examples{
\dontrun{
  dat<-read.ms.snps(system.file("extdata","2pops_asspop.txt",package="hierfstat"))
  betas(dat)$betaiovl # population specif FSTs
  }
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
\references{
\href{http://bioinformatics.oxfordjournals.org/content/18/2/337.short}{Hudson, R. R. (2002) Generating samples under a Wright-Fisher neutral model of genetic variation}. Bioinformatics 18 : 337-338.
}

