% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_dosage.R
\name{TajimaD.dosage}
\alias{TajimaD.dosage}
\title{Estimates Tajima's D}
\usage{
TajimaD.dosage(dos)
}
\arguments{
\item{dos}{a ni X nl dosage matrix containing the number of derived/alternate alleles each individual carries
at each SNP}
}
\value{
Tajima's D (eqn 38 of Tajima, 1989)
}
\description{
Estimates Tajima's D from dosage data
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1203831/}{Tajima F. 1989} 
Statistical Method for Testing the Neutral Mutation Hypothesis by DNA Polymorphism. Genetics 123:585-595.
}
