% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_dosage.R
\name{theta.Watt.dosage}
\alias{theta.Watt.dosage}
\title{Estimates \eqn{\theta_{Watterson}} from dosage data}
\usage{
theta.Watt.dosage(dos,L=NULL)
}
\arguments{
\item{dos}{a ni X nl dosage matrix containing the number of derived/alternate alleles each individual carries
at each SNP}

\item{L}{the length of the sequence}
}
\value{
if \code{L=NULL} (default), returns \eqn{\theta_{Watterson}}, else return \eqn{\theta_{Watterson}/L}
}
\description{
Estimates \eqn{\theta_{Watterson}=S/a}, where \eqn{S} is the number of segregating sites
in a set of sequences and \eqn{a=1/\sum_i^{n-1} i}.
}
