\name{vec2mat}
\alias{vec2mat}
\title{Reads a vector into a matrix}
\description{Fills a lower triangular matrix from a vector and copy it to upper triangle}
\usage{vec2mat(x)}
\arguments{
\item{x}{a vector}
}

\value{
\item{mat}{a matrix}
} 
%\references{}
\author{Jerome Goudet \email{jerome.goudet@unil.ch}}

%\seealso{\code{\link{}}.}
%\examples{vec2mat(1:6)}
\keyword{univar}
