% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writebayescan.R
\name{write.bayescan}
\alias{write.bayescan}
\title{write.bayescan Writes a bayescan file}
\usage{
write.bayescan(dat=dat,diploid=TRUE,fn="dat.bsc")
}
\arguments{
\item{dat}{a genotype data frame}

\item{diploid}{whether the dataset is diploid or haplois}

\item{fn}{file name for output}
}
\value{
a text file fn is written in the current directory
}
\description{
write the genotypes in a format suitable for
analysis with bayescan
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
\references{
\href{Foll M and OE Gaggiotti (2008) A genome scan method to identify
selected loci appropriate for both dominant and codominant markers:
A Bayesian perspective. Genetics 180: 977-993}
{http://www.genetics.org/content/180/2/977.abstract}

\url{http://cmpg.unibe.ch/software/BayeScan/}
}

