% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat.from.adegenet.R
\name{fstat}
\alias{fstat}
\alias{FST}
\alias{fst}
\title{Wrapper for fst estimator from hierfstat package (from adegenet)}
\usage{
fstat(x, pop = NULL, fstonly = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{genind}.}

\item{pop}{a factor giving the 'population' of each individual. If NULL, pop
is seeked from \code{pop(x)}. Note that the term population refers in fact
to any grouping of individuals'.}

\item{fstonly}{a logical stating whether only the Fst should be returned.}
}
\value{
A vector, a matrix, or a dist object containing F statistics.
}
\description{
The function \code{fstat} is a wrapper for \code{varcomp.glob} of the
package \code{hierfstat}. For Fst, Fis and Fit, an alternative is offered by
\code{Fst} from the \code{pegas} package (see example).
}
\details{
Let \eqn{A} and \eqn{B} be two populations of population sizes \eqn{n_A} and
\eqn{n_B}, with expected heterozygosity (averaged over loci) \eqn{Hs(A)} and
\eqn{Hs(B)}, respectively. We denote \eqn{Ht} the expected heterozygosity of
a population pooling \eqn{A} and \eqn{B}. Then, the pairwise \eqn{Fst}
between \eqn{A} and \eqn{B} is computed as:\cr

\eqn{ Fst(A,B) = \frac{(Ht - (n_A Hs(A) + n_B Hs(B))/(n_A + n_B) )}{Ht}} \cr
}
\examples{

\dontrun{
if(require(adegenet)){
data(nancycats)


## Fst, Fis, Fit
## using hierfstat
if(require(hierfstat)){
fstat(nancycats)
}

## using pegas
if(require(pegas)){
data(nancycats)

## conversion to pegas's format
as.loci(nancycats)

## use Fst from pegas
fsttab <- Fst(as.loci(nancycats))

## average over loci
apply(fsttab, 2, mean)
}
}
}
}
\references{
Nei, M. (1973) Analysis of gene diversity in subdivided
populations. Proc Natl Acad Sci USA, 70: 3321-3323
}
\seealso{
\code{\link[adegenet]{Hs}}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}

Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
\keyword{multivariate}
