\name{pp.fst}
\alias{pp.fst}
\title{fst per pair}
\description{fst per pair}
\usage{pp.fst(dat=dat,diploid=TRUE,...)}
\arguments{
\item{dat}{a genetic data frame}
\item{diploid}{whether data from diploid organism}
\item{...}{further arguments to pass to the function}
}

\value{
\item{call}{function call}
\item{fst.pp}{pairwise Fsts}
\item{vc.per.loc}{for each pair of population, the variance components per locus}
}
\author{Jerome Goudet \email{jerome.goudet@unil.ch}}

%\seealso{\code{\link{}}.}
%\examples{}
\keyword{univar}
