% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2mat.R
\name{vec2mat}
\alias{vec2mat}
\title{Fills a triangular matrix from the inputed vector}
\usage{
vec2mat(x,diag=FALSE,upper=FALSE)
}
\arguments{
\item{x}{a vector}

\item{diag}{whether the vector contains the diagonal elements}

\item{upper}{whether the vector contains the upper trinagular matrix elements}
}
\value{
a matrix
}
\description{
Fills a triangular matrix from the inputed vector
}
\examples{
{
 vec2mat(1:10)
 vec2mat(1:10,diag=TRUE)
 vec2mat(1:10,upper=TRUE)
}
}
