% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeped.R
\name{write.ped}
\alias{write.ped}
\title{Write ped file plink analysis}
\usage{
write.ped(dat, ilab = NULL, pop = NULL, 
        fname = "dat",na.str="0",f.id=NULL,m.id=NULL,loc.pos=NULL,sex=NULL)
}
\arguments{
\item{dat}{a hierfstat data frame}

\item{ilab}{individal labels}

\item{pop}{population id}

\item{fname}{filename for ped file}

\item{na.str}{character string to use for missing values}

\item{f.id}{father id. default to unknown}

\item{m.id}{mother id. default to unknown}

\item{loc.pos}{the loci position default to unknown}

\item{sex}{the individual sex. default to unknown}
}
\value{
a map file containing the loci positions

a ped file containing genotypes etc...
}
\description{
write a ped and a map file suitable for analysis with 
\href{plink}{http://pngu.mgh.harvard.edu/~purcell/plink/}
}
\references{
\href{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC1950838/}{Purcell etal (2007) PLINK:} A Tool Set for Whole-Genome Association 
 and Population-Based Linkage Analyses 81:559-575
}

