% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeped.R
\name{write.ped}
\alias{write.ped}
\title{Write ped file plink analysis}
\usage{
write.ped(dat, ilab = NULL, pop = NULL,
        fname = "dat",na.str="0",f.id=NULL,m.id=NULL,loc.pos=NULL)
}
\arguments{
\item{dat}{}

\item{ilab}{}

\item{pop}{}

\item{fname}{}

\item{na.str}{}

\item{f.id}{}

\item{m.id}{}

\item{loc.pos}{}
}
\value{
a map file containing the loci positions

a ped file containing genotypes etc...
}
\description{
write a ped and a map file suitable for analysis with
\href{plink}{http://pngu.mgh.harvard.edu/~purcell/plink/}
}
\references{
\href{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC1950838/}{Purcell etal (2007) PLINK:} A Tool Set for Whole-Genome Association
 and Population-Based Linkage Analyses 81:559-575
}

